feat(os): add freebsd support

+++ salter.sh
@@ -28,7 +28,7 @@ RC=0
 BASE=/srv
 BASE_ETC=/etc
 STATEDIR=''
-if [ `uname` == 'FreeBSD' ]; then
+if [ `uname` == "FreeBSD" ]; then
     BASE=/usr/local/etc
     BASE_ETC=/usr/local/etc
     STATEDIR=states
@@ -70,6 +70,8 @@ pkg-query() {
          /usr/bin/dpkg-query --list | grep ${PACKAGE}
     elif [ -f "/usr/bin/pacman" ]; then
          /usr/bin/pacman -Qi ${PACKAGE}
+    elif [[ -f "/usr/sbin/pkg" ]]; then
+         /usr/sbin/pkg query ${PACKAGE}
     fi
 }
 
@@ -91,7 +93,8 @@ pkg-install() {
              fi
              ;;
 
-    linux*)  if [ -f "/usr/bin/zypper" ]; then
+    linux*|freebsd*)
+             if [ -f "/usr/bin/zypper" ]; then
                  /usr/bin/zypper update -y || exit 1
                  /usr/bin/zypper --non-interactive install ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/emerge" ]; then
@@ -103,14 +106,17 @@ pkg-install() {
              elif [ -f "/usr/bin/dnf" ]; then
                  /usr/bin/dnf install -y --best --allowerasing ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/yum" ]; then
-                 /usr/bin/yum update -y || exit 1 
-                 /usr/bin/yum install -y ${PACKAGES} || exit 1 
+                 /usr/bin/yum update -y || exit 1
+                 /usr/bin/yum install -y ${PACKAGES} || exit 1
              elif [[ -f "/usr/bin/apt-get" ]]; then
                  /usr/bin/apt-get update --fix-missing -y || exit 1
                  /usr/bin/apt-add-repository universe
                  /usr/bin/apt autoremove -y
                  /usr/bin/apt-get update -y
                  /usr/bin/apt-get install -y ${PACKAGES} || exit 1
+             elif [[ -f "/usr/sbin/pkg" ]]; then
+                 /usr/sbin/pkg update -f --quiet || exit 1
+                 /usr/sbin/pkg install --automatic --yes ${PACKAGES} || exit 1
              fi
     esac
 }
@@ -131,9 +137,11 @@ pkg-update() {
              elif [ -f "/usr/bin/dnf" ]; then
                  /usr/bin/dnf upgrade -y --allowerasing ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/yum" ]; then
-                 /usr/bin/yum update -y ${PACKAGES} || exit 1 
+                 /usr/bin/yum update -y ${PACKAGES} || exit 1
              elif [[ -f "/usr/bin/apt-get" ]]; then
                  /usr/bin/apt-get upgrade -y ${PACKAGES} || exit 1
+             elif [[ -f "/usr/sbin/pkg" ]]; then
+                 /usr/sbin/pkg upgrade --yes ${PACKAGES} || exit 1
              fi
     esac
 }
@@ -145,7 +153,8 @@ pkg-remove() {
                  su - ${USER} -c "${HOMEBREW} uninstall ${p} --force"
              done
              ;;
-    linux*)  if [ -f "/usr/bin/zypper" ]; then
+    linux*|freebsd*)
+             if [ -f "/usr/bin/zypper" ]; then
                  /usr/bin/zypper --non-interactive rm ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/emerge" ]; then
                  /usr/bin/emerge -C ${PACKAGES} || exit 1
@@ -154,9 +163,11 @@ pkg-remove() {
              elif [ -f "/usr/bin/dnf" ]; then
                  /usr/bin/dnf remove -y ${PACKAGES} || exit 1
              elif [ -f "/usr/bin/yum" ]; then
-                 /usr/bin/yum remove -y ${PACKAGES} || exit 1 
+                 /usr/bin/yum remove -y ${PACKAGES} || exit 1
              elif [[ -f "/usr/bin/apt-get" ]]; then
                  /usr/bin/apt-get remove -y ${PACKAGES} || exit 1
+             elif [[ -f "/usr/sbin/pkg" ]]; then
+                 /usr/sbin/pkg delete --yes ${PACKAGES} || exit 1
              fi
     esac
 }
@@ -228,8 +239,9 @@ salt-bootstrap() {
              export HOMEBREW_CURLRC=1
              ;;
 
-     linux*) pkg-update 2>/dev/null
-             echo "Setup Linux baseline and install saltstack masterless minion ..."
+     linux*|freebsd*)
+             pkg-update 2>/dev/null
+             echo "Setup Linux/FreeBSD baseline and install saltstack masterless minion ..."
              if [ -f "/usr/bin/dnf" ]; then
                  PACKAGES="--best --allowerasing git wget redhat-rpm-config"
              elif [ -f "/usr/bin/yum" ]; then
@@ -240,6 +252,8 @@ salt-bootstrap() {
                  PACKAGES="git ssh wget curl software-properties-common"
              elif [ -f "/usr/bin/pacman" ]; then
                  PACKAGES="git wget psutils"
+             elif [ -f "/usr/sbin/pkg" ]; then
+                 PACKAGES="git wget psutils"
              fi
              pkg-install ${PACKAGES} 2>/dev/null
              if (( $? > 0 )); then
@@ -249,6 +263,7 @@ salt-bootstrap() {
              wget -O install_salt.sh https://bootstrap.saltstack.com || exit 10
              (sh install_salt.sh -x python3 ${SALT_VERSION} && rm -f install_salt.sh) || exit 10
              rm -f install_salt.sh 2>/dev/null
+             ;;
     esac
     ### stop debian interference with services (https://wiki.debian.org/chroot)
     if [ -f "/usr/bin/apt-get" ]; then
@@ -260,22 +275,30 @@ EOF
         ### Enforce python3
         rm /usr/bin/python 2>/dev/null; ln -s /usr/bin/python3 /usr/bin/python
     fi
-    ### install salt-api (except arch/macos)
-    [ ! -f "/etc/arch-release" ] && [ "$(uname)" != "Darwin" ] && pkg-install salt-api
+    ### install salt-api (except arch/macos/freebsd)
+    [ ! -f "/etc/arch-release" ] && [ "$(uname)" != "Darwin" ] && [ "$(uname)" != "FreeBSD" ] && pkg-install salt-api
 
-    ### salt services
+    ### salt minion
+    [ ! -f "${BASE_ETC}/salt/minion" ] && echo "File ${BASE_ETC}/salt/minion not found" && exit 1
     if [[ "`uname`" == "FreeBSD" ]] || [[ "`uname`" == "Darwin" ]]; then
         sed -i"bak" "s@^\s*#*\s*master\s*: salt\s*\$@master: ${solution[saltmaster]}@" ${BASE_ETC}/salt/minion
     else
         sed -i "s@^\s*#*\s*master\s*: salt\s*\$@master: ${solution[saltmaster]}@" ${BASE_ETC}/salt/minion
     fi
+    ### salt services
     (systemctl enable salt-api && systemctl start salt-api) 2>/dev/null || service start salt-api 2>/dev/null
     (systemctl enable salt-master && systemctl start salt-master) 2>/dev/null || service start salt-master 2>/dev/null
     (systemctl enable salt-minion && systemctl start salt-minion) 2>/dev/null || service start salt-minion 2>/dev/null
     salt-key -A --yes >/dev/null 2>&1     ##accept pending registrations
+
+    ### reboot to activate a new kernel?
     echo && KERNEL_VERSION=$( uname -r | awk -F. '{print $1"."$2"."$3"."$4"."$5}' )
     echo "kernel before: ${KERNEL_VERSION}"
-    echo "kernel after: $( pkg-query linux 2>/dev/null )"
+    if [ "$(uname)" == "FreeBSD" ]; then
+        echo "kernel after: $( /bin/freebsd-version -k 2>/dev/null )"
+    else
+        echo "kernel after: $( pkg-query linux 2>/dev/null )"
+    fi
     echo "Reboot if kernel was major-upgraded; if unsure reboot!"
     echo
 }
@@ -315,7 +338,7 @@ highstate() {
     ## prepare states
     ACTION=${1} && STATEDIR=${2} && TARGET=${3}
     for PROFILE in ${solution[saltdir]}/${ACTION}/${TARGET} ${your[saltdir]}/${ACTION}/${TARGET}
-    do  
+    do
         [ -f ${PROFILE}.sls ] && cp ${PROFILE}.sls ${SALTFS}/top.sls && break
         [ -f ${PROFILE}/init.sls ] && cp ${PROFILE}/init.sls ${SALTFS}/top.sls && break
     done
@@ -366,7 +389,7 @@ usage() {
     echo "  [-u <username>]" 1>&2
     echo "        A Loginname (current or corporate or root user)." 1>&2
     echo "        its mandatory for some Linux profiles" 1>&2
-    echo "        but not required on MacOS" 1>&2 
+    echo "        but not required on MacOS" 1>&2
     echo 1>&2
     echo "  TARGETS" 1>&2
     echo 1>&2
